Rename Command Table
